requirements.txt
streamlit==1.32.0
google-generativeai==0.4.1
requests==2.31.0
psycopg2-binary==2.9.9
pgvector==0.2.5